Fix mastra create stater workflow breaking with model provider not ai-sdk/openai
