+++ packages/cli/package.json
@@ -74,7 +74,6 @@
     "zod-to-json-schema": "^3.24.1"
   },
   "devDependencies": {
-    "@ai-sdk/openai": "^1.1.9",
     "@internal/lint": "workspace:*",
     "@microsoft/api-extractor": "^7.49.2",
     "@types/express": "^5.0.0",
